refactor(e2e): separate API smoke tests from UI tests, add trace for videos
- Move 19 API-only smoke specs to tests/e2e/smoke/ (no UI interactions)
- Add "smoke" project in playwright.config.ts (no video, no screenshot)
- Keep "chromium" project for real UI tests only (with trace: "on" for click tracking)
- Fix relative imports in smoke specs (./helpers/auth → ../helpers/auth)

Projects:
  chromium — 188 UI tests (video + trace for documentation)
  smoke    — 89 API contract tests (fast, no video)
  scenarios — 12 Documentation Vivante (human-paced, YouTube)

All green locally: 277 passed, 0 failed, 21 skipped
