+++ frontend/playwright.config.ts
@@ -63,8 +63,27 @@ export default defineConfig({
         ...devices["Desktop Chrome"],
         viewport: { width: 1280, height: 720 },
         locale: "fr-BE",
+        trace: "on",
       },
-      testIgnore: /scenarios\//,
+      testIgnore: [/scenarios\//, /smoke\//],
+    },
+
+    /**
+     * API smoke tests — no video, parallel workers, fast.
+     * These test backend API contracts, not UI interactions.
+     *
+     * Run only smokes:  npx playwright test --project=smoke
+     */
+    {
+      name: "smoke",
+      testDir: "./tests/e2e/smoke",
+      use: {
+        ...devices["Desktop Chrome"],
+        viewport: { width: 1280, height: 720 },
+        locale: "fr-BE",
+        video: "off",
+        screenshot: "off",
+      },
     },
 
     /**
