+++ frontend/tests/e2e/smoke/Gamification.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import { loginAsSyndic } from "./helpers/auth";
+import { loginAsSyndic } from "../helpers/auth";
 
 /**
  * Gamification E2E Test Suite - Achievements & Challenges
