+++ frontend/tests/e2e/smoke/Marketplace.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import { loginAsSyndic } from "./helpers/auth";
+import { loginAsSyndic } from "../helpers/auth";
 
 /**
  * Marketplace E2E Test Suite - Service Provider Directory
