+++ frontend/tests/e2e/smoke/OwnerContributions.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import { loginAsSyndicWithOwner } from "./helpers/auth";
+import { loginAsSyndicWithOwner } from "../helpers/auth";
 
 const API_BASE = process.env.PLAYWRIGHT_API_BASE || "http://localhost/api/v1";
