+++ frontend/tests/e2e/smoke/PublicSyndic.spec.ts
