+++ frontend/tests/e2e/smoke/TwoFactor.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import { loginAsSyndic } from "./helpers/auth";
+import { loginAsSyndic } from "../helpers/auth";
 
 /**
  * Two-Factor Authentication E2E Test Suite - TOTP 2FA
