+++ frontend/tests/e2e/smoke/UnitOwners.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import { loginAsSyndicWithUnit } from "./helpers/auth";
+import { loginAsSyndicWithUnit } from "../helpers/auth";
 
 const API_BASE = process.env.PLAYWRIGHT_API_BASE || "http://localhost/api/v1";
